extract gamedata into a separate file.
